objectos revolucao e extrusao melhores.

+++ VertexArray.cpp
@@ -17,7 +17,7 @@ void VertexArray::AddBuffer(const VertexBuffer &vb,
   Bind();
   vb.Bind();
   const auto &elements = layout.GetElements();
-  unsigned int offset = 0;
+  unsigned long int offset = 0;
   for (unsigned int i = 0; i < elements.size(); i++) {
     const auto &element = elements[i];
     glVertexAttribPointer(i, element.count, element.type, element.normalized,
